- Some cleanups - Implemented vi2s and vi2us - Fixed game list GUI synchronization - Fixed GpuStateStruct size (2048) and reduced base size - Support for scissors - Implemented sceGeSaveContext and sceGeRestoreContext - Implemented sceAtracGetRemainFrame - Implemented scePowerRegisterCallback - Fixed sceKernelPowerLock - Fixed sceKernelDelayThread* now Astonishia Story works again - Fixed MemoryStick :: CommandType.CheckMemoryStickStatus now lumines detects the memory stick (and probably other games) - Added readonly to some static fields - More random work

+++ CSPspEmu.Core.Gpu/Run/GpuDisplayListRunner.Special.cs
@@ -129,8 +129,8 @@ namespace CSPspEmu.Core.Gpu.Run
 		/// </summary>
 		public void OP_SCISSOR1()
 		{
-			GpuState->ClipPlaneState.Scissor.Left = BitUtils.Extract(Params24, 0, 10);
-			GpuState->ClipPlaneState.Scissor.Top = BitUtils.Extract(Params24, 10, 10);
+			GpuState->ClipPlaneState.Scissor.Left = (short)BitUtils.Extract(Params24, 0, 10);
+			GpuState->ClipPlaneState.Scissor.Top = (short)BitUtils.Extract(Params24, 10, 10);
 		}
 
 		/// <summary>
@@ -138,8 +138,8 @@ namespace CSPspEmu.Core.Gpu.Run
 		/// </summary>
 		public void OP_SCISSOR2()
 		{
-			GpuState->ClipPlaneState.Scissor.Right = BitUtils.Extract(Params24, 0, 10);
-			GpuState->ClipPlaneState.Scissor.Bottom = BitUtils.Extract(Params24, 10, 10);
+			GpuState->ClipPlaneState.Scissor.Right = (short)BitUtils.Extract(Params24, 0, 10);
+			GpuState->ClipPlaneState.Scissor.Bottom = (short)BitUtils.Extract(Params24, 10, 10);
 		}
 
 		/**
